Pendente apenas "teste de conectividade" do modo LOGGER.

+++ logger_DatagramRead.m
@@ -2,7 +2,7 @@ function outInfo = logger_DatagramRead(DataType, Datagram, opt)
 
     arguments
         DataType int32
-        Datagram uint8
+        Datagram double
         opt      uint32 = []
     end
 
@@ -27,11 +27,11 @@ function GPS = Fcn_GPSRead(Datagram)
     GPS.TimeStamp        = datestr(datetime([Date_Year, Date_Month, Date_Day, Time_Hours, Time_Minutes, Time_Seconds]), 'dd/mm/yyyy HH:MM:ss');
     GPS.Status           = Datagram(41);
 %   GPS.SatellitesInView = Datagram(42);
-%   GPS.Heading          = double(typecast(Datagram(43:44), 'uint16')) / 100;
-    GPS.Latitude         = double(typecast(Datagram(45:48), 'int32' )) / 1e+6;
-    GPS.Longitude        = double(typecast(Datagram(49:52), 'int32' )) / 1e+6;
-%   GPS.Speed            = double(typecast(Datagram(53:56), 'uint32')) / 1000;
-    GPS.Altitude         = double(typecast(Datagram(57:60), 'uint32')) / 1000;
+%   GPS.Heading          = double(typecast(uint8(Datagram(43:44)), 'uint16')) / 100;
+    GPS.Latitude         = double(typecast(uint8(Datagram(45:48)), 'int32' )) / 1e+6;
+    GPS.Longitude        = double(typecast(uint8(Datagram(49:52)), 'int32' )) / 1e+6;
+%   GPS.Speed            = double(typecast(uint8(Datagram(53:56)), 'uint32')) / 1000;
+    GPS.Altitude         = double(typecast(uint8(Datagram(57:60)), 'uint32')) / 1000;
     
 end
 
@@ -47,7 +47,7 @@ function outInfo = Fcn_TextRead(Datagram)
     TimeStamp    = datestr(datetime([Date_Year, Date_Month, Date_Day, Time_Hours, Time_Minutes, Time_Seconds]), 'dd/mm/yyyy HH:MM:ss');
 
     Identifier     = deblank(char(Datagram(41:72)));
-    FreeTextLength = double(typecast(Datagram(73:76), 'uint32'));
+    FreeTextLength = double(typecast(uint8(Datagram(73:76)), 'uint32'));
     if FreeTextLength
         FreeText = deblank(char(Datagram(77:76+FreeTextLength)));
     end
@@ -70,18 +70,18 @@ function outInfo = Fcn_SpectralRead(DataType, Datagram, ThreadID)
                  'StepWidth',   [], ...
                  'REC',         []);
     
-    DescriptionLength = double(typecast(Datagram(41:44), 'uint32'));
+    DescriptionLength = double(typecast(uint8(Datagram(41:44)), 'uint32'));
     Bin.Description   = deblank(char(Datagram(45:45+DescriptionLength-1)));
     
-    IntegerPart    = double(typecast(Datagram(45+DescriptionLength:46+DescriptionLength), 'uint16'));
-    DecimalPart    = double(typecast(Datagram(47+DescriptionLength:50+DescriptionLength), 'int32'));
+    IntegerPart    = double(typecast(uint8(Datagram(45+DescriptionLength:46+DescriptionLength)), 'uint16'));
+    DecimalPart    = double(typecast(uint8(Datagram(47+DescriptionLength:50+DescriptionLength)), 'int32'));
     Bin.FreqStart  = (IntegerPart + DecimalPart./1e+9) .* 1e+6;
     
-    IntegerPart    = double(typecast(Datagram(51+DescriptionLength:52+DescriptionLength), 'uint16'));
-    DecimalPart    = double(typecast(Datagram(53+DescriptionLength:56+DescriptionLength), 'int32'));
+    IntegerPart    = double(typecast(uint8(Datagram(51+DescriptionLength:52+DescriptionLength)), 'uint16'));
+    DecimalPart    = double(typecast(uint8(Datagram(53+DescriptionLength:56+DescriptionLength)), 'int32'));
     Bin.FreqStop   = (IntegerPart + DecimalPart./1e+9) .* 1e+6;
     
-    Bin.Resolution = double(typecast(Datagram(57+DescriptionLength:60+DescriptionLength), 'uint32'));
+    Bin.Resolution = double(typecast(uint8(Datagram(57+DescriptionLength:60+DescriptionLength)), 'uint32'));
     
     switch Datagram(74+DescriptionLength)
         case 0; Bin.Operation = 'Single Measurement';
@@ -95,22 +95,22 @@ function outInfo = Fcn_SpectralRead(DataType, Datagram, ThreadID)
         case 1; Bin.LevelUnit = 'dBµV/m';
     end
     
-    OFFSET = double(typecast(Datagram(76+DescriptionLength), 'int8'));
-    NTUN   = typecast(Datagram(80+DescriptionLength:81+DescriptionLength), 'uint16');
-    NAGC   = typecast(Datagram(82+DescriptionLength:83+DescriptionLength), 'uint16');
+    OFFSET = double(typecast(uint8(Datagram(76+DescriptionLength)), 'int8'));
+    NTUN   = typecast(uint8(Datagram(80+DescriptionLength:81+DescriptionLength)), 'uint16');
+    NAGC   = typecast(uint8(Datagram(82+DescriptionLength:83+DescriptionLength)), 'uint16');
 
     switch DataType
         case 67
-            Bin.DataPoints = double(typecast(Datagram(85+DescriptionLength:88+DescriptionLength), 'uint32'));
+            Bin.DataPoints = double(typecast(uint8(Datagram(85+DescriptionLength:88+DescriptionLength)), 'uint32'));
             Bin.StepWidth  = (Bin.FreqStop-Bin.FreqStart)/(Bin.DataPoints-1);
 
             newArray = Datagram(89+DescriptionLength+4*NTUN+NAGC:89+DescriptionLength+4*NTUN+NAGC+Bin.DataPoints-1)./2 + OFFSET - 127.5;
 
         case 68
-            NCDATA = double(typecast(Datagram(85+DescriptionLength:88+DescriptionLength), 'uint32'));
-            THRESH = double(typecast(Datagram(89+DescriptionLength:92+DescriptionLength), 'int32'));
+            NCDATA = double(typecast(uint8(Datagram(85+DescriptionLength:88+DescriptionLength)), 'uint32'));
+            THRESH = double(typecast(uint8(Datagram(89+DescriptionLength:92+DescriptionLength)), 'int32'));
         
-            Bin.DataPoints = double(typecast(Datagram(93+DescriptionLength:96+DescriptionLength), 'uint32'));
+            Bin.DataPoints = double(typecast(uint8(Datagram(93+DescriptionLength:96+DescriptionLength)), 'uint32'));
             Bin.StepWidth  = (Bin.FreqStop-Bin.FreqStart)/(Bin.DataPoints-1);
         
             CompressedData = Datagram(97+DescriptionLength+4*NTUN+NAGC:97+DescriptionLength+4*NTUN+NAGC+NCDATA-1);
